server/requirements.txt
fastapi
uvicorn[standard]
requests
pillow
pydantic
python-multipart
python-json-logger